+++ app/src/main/java/com/snoffee/app/data/model/SleepDataDto.kt
@@ -8,5 +8,5 @@ data class SleepDataDto(
     val date: LocalDateTime,                // 수면 날짜 (PK)
     val sleepStart: Long,          // 수면 시작 시각
     val sleepEnd: Long,            // 수면 종료 시각
-    val deepSleepRatio: Double              // 깊은 수면 비율 (0.0 ~ 1.0)
+    val deepSleepRatio: Int             // 깊은 수면 비율 (0.0 ~ 1.0)
 )
